+++ r2c_isg/loaders/core.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from r2c_isg.structures import Dataset
+from r2c_isg.dataset import Dataset
 
 
 class Loader(ABC):
