--- r2c_isg/loaders/core/__init__.py
@@ -1,3 +1,0 @@
-from .dataset_loader import DatasetLoader
-from .r2c_loader import R2cLoader
-
