+++ r2c_isg/loaders/dataset_loader.py
@@ -1,7 +1,7 @@
 import dill as pickle
 
-from r2c_isg.loaders import Loader
-from r2c_isg.structures import Dataset
+from r2c_isg.loaders.core import Loader
+from r2c_isg.dataset import Dataset
 
 
 class DatasetLoader(Loader):
