+++ r2c_isg/loaders/file/csv_loader.py
@@ -1,7 +1,8 @@
 import csv
 
 from r2c_isg.loaders import Loader
-from r2c_isg.structures import Dataset, DefaultProject, DefaultVersion
+from r2c_isg.structures import DefaultProject, DefaultVersion
+from r2c_isg.dataset import Dataset
 from r2c_isg.structures.projects import project_map
 from r2c_isg.structures.versions import version_map
