+++ r2c_isg/loaders/r2c_loader.py
@@ -2,7 +2,8 @@ import json
 from tqdm import tqdm
 
 from r2c_isg.loaders import Loader
-from r2c_isg.structures import Dataset, DefaultProject, DefaultVersion
+from r2c_isg.structures import DefaultProject, DefaultVersion
+from r2c_isg.dataset import Dataset
 from r2c_isg.structures.projects import project_map
 from r2c_isg.structures.versions import version_map
